this is better….

+++ kernel/dev/ramdisk/ramdisk.c
@@ -40,7 +40,7 @@ uint64_t ramdisk_calc_address(uint64_t data, uint32_t sector){
     return data+(sector*RAMDISK_SECTOR_SIZE);
 }
 
-void ramdisk_read(struct device* dev, uint32_t sector, uint16_t* data, uint32_t count) {
+void ramdisk_read(struct device* dev, uint32_t sector, uint8_t* data, uint32_t count) {
 	ASSERT_NOT_NULL(dev, "dev cannot be null");
 	ASSERT_NOT_NULL(data, "data cannot be null");
     ASSERT_NOT_NULL(dev->deviceData, "dev->deviceData cannot be null");
@@ -49,10 +49,10 @@ void ramdisk_read(struct device* dev, uint32_t sector, uint16_t* data, uint32_t 
     ASSERT(sector < RAMDISK_TOTAL_SECTORS, "invalid sector");
     uint64_t block = ramdisk_calc_address((uint64_t)deviceData->data, sector);
     kprintf("block %#hX\n", block);
-    memcpy((uint8_t*)data, (uint8_t*)block, count*sizeof(uint16_t));
+    memcpy((uint8_t*)data, (uint8_t*)block, count*sizeof(uint8_t));
 }
 
-void ramdisk_write(struct device* dev, uint32_t sector, uint16_t* data, uint32_t count) {
+void ramdisk_write(struct device* dev, uint32_t sector, uint8_t* data, uint32_t count) {
 	ASSERT_NOT_NULL(dev, "dev cannot be null");
 	ASSERT_NOT_NULL(data, "data cannot be null");
     ASSERT_NOT_NULL(dev->deviceData, "dev->deviceData cannot be null");
@@ -61,7 +61,7 @@ void ramdisk_write(struct device* dev, uint32_t sector, uint16_t* data, uint32_t
     ASSERT(sector < RAMDISK_TOTAL_SECTORS, "invalid sector");
     uint64_t block = ramdisk_calc_address((uint64_t)deviceData->data, sector);
     kprintf("block %#hX\n", block);
-    memcpy((uint8_t*)block, (uint8_t*)data, count*sizeof(uint16_t));  
+    memcpy((uint8_t*)block, (uint8_t*)data, count*sizeof(uint8_t));  
 }
 
 uint16_t ramdisk_sector_size(struct device* dev) {
